requirements.txt
pillow
jpype
numpy
skimage
matplotlib
pandas
opencv-python